Bug fixed
When you kill a process and calling Wox again, it shows the process if you don't type anything and press ENTER, letting you kill a non existant process throwing an error.
I've added a check for HasExited in every p.kill
The ideal would be that the list clears whan you press ENTER but i don't know how to do it :)

+++ Main.cs
@@ -77,7 +77,11 @@ namespace Wox.Plugin.ProcessKiller
                         SubTitle = path,
                         Action = (c) =>
                         {
-                            p.Kill();
+                            if (!p.HasExited)
+                            {
+                                p.Kill();
+                            }
+                            
                             return true;
                         }
                     });
@@ -95,7 +99,10 @@ namespace Wox.Plugin.ProcessKiller
                            {
                                foreach (var p in processlist)
                                {
-                                   p.Kill();
+                                   if (!p.HasExited)
+                                   {
+                                        p.Kill();
+                                   }
                                }
                                return true;
                            }
